requirements.txt
azure.identity
azure-mgmt-apimanagement
requests
urllib3==1.26.6